requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
Flask-Admin
mysqlclient
python-dotenv
requests